Renaming sparse_algorithm|parameters
