+++ gp/src/lib.rs
@@ -14,14 +14,14 @@ mod algorithm;
 pub mod correlation_models;
 mod errors;
 pub mod mean_models;
-mod sgp_algorithm;
+mod sparse_algorithm;
 
 mod parameters;
-mod sgp_parameters;
+mod sparse_parameters;
 mod utils;
 
 pub use algorithm::*;
 pub use errors::*;
 pub use parameters::*;
-pub use sgp_algorithm::*;
-pub use sgp_parameters::*;
+pub use sparse_algorithm::*;
+pub use sparse_parameters::*;
